906: Moved getOrdersFromLogs to transaction helpers
